// src/app/shared/models/enum.ts
export enum OrderStatus {
    Created = 'Created',
    PendingApproval = 'PendingApproval',
    PendingPayment = 'PendingPayment',
    SubmittedToWarehouse = 'SubmittedToWarehouse',
    BeingPacked = 'BeingPacked',
    Shipped = 'Shipped',
    Delivered = 'Delivered',
    PaymentDeclined = 'PaymentDeclined',
  }